% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_from_image.R
\name{area_from_image}
\alias{area_from_image}
\title{Get Soil Area for a Single Image}
\usage{
area_from_image(
  path_to_rgb,
  interactive = FALSE,
  circular_mask = TRUE,
  false_color = "red",
  d = 0.7,
  h_offset = 0,
  v_offset = 0,
  w_new = NULL,
  h_new = NULL,
  aggregates_in_initial = NULL,
  max_rel_dist_from_centroid = 1.3,
  max_rel_dist_from_center = 0.37,
  fixed_crop_fraction = NULL,
  max_rel_objlength = 0.3,
  automask_buffer = 0.15,
  erode_kern = 11,
  dilate_kern = 31,
  normdiff_min = 0
)
}
\arguments{
\item{path_to_rgb}{character string containing relative or absolute path to color image}

\item{interactive}{logical. If \code{TRUE}, image masking will be conducted interactively.}

\item{circular_mask}{logical. If \code{TRUE}, a circular mask will be applied.}

\item{false_color}{character string with color name or hex code for false color overlay}

\item{d}{diameter of circular mask as a proportion of image width}

\item{h_offset}{horizontal offset for center of circular mask}

\item{v_offset}{vertical offset for center of circular mask}

\item{w_new}{width for image resizing}

\item{h_new}{height for image resizing}

\item{aggregates_in_initial}{number of soil aggregates in initial (pre-submersion) image. If not NULL (default), this number of the largest aggregates will be retained, and all smaller objects classified as soil will be disregarded.}

\item{max_rel_dist_from_centroid}{maximum allowed relative distance from centroid of largest three objects}

\item{max_rel_dist_from_center}{maximum allowed relative distance from image center}

\item{fixed_crop_fraction}{proportion of image width to crop. If not NULL (default), must be between zero and one (exclusive).}

\item{max_rel_objlength}{maximum allowed object length as a proportion of image width}

\item{automask_buffer}{proportional buffer for automated masking. Only used if interactive=FALSE, circular_mask=FALSE, and fixed_crop_fraction=NULL.}

\item{erode_kern}{erosion kernel size for automated masking. Only used if interactive=FALSE, circular_mask=FALSE, and fixed_crop_fraction=NULL. See \code{\link[EBImage:morphology]{EBImage::erode()}}.}

\item{dilate_kern}{dilation kernel size for automated masking. Only used if interactive=FALSE, circular_mask=FALSE, and fixed_crop_fraction=NULL. See \code{\link[EBImage:morphology]{EBImage::dilate()}}.}

\item{normdiff_min}{minimum value of normalized difference between red and green bands for automated masking. Only used if interactive=FALSE, circular_mask=FALSE, and fixed_crop_fraction=NULL. See \code{\link[=norm_diff]{norm_diff()}}.}
}
\value{
A three-element list, containing the area of soil (in pixels), an \code{Image} object with labeled soil pixels (see \code{\link[EBImage:bwlabel]{EBImage::bwlabel()}}), and the original \code{Image} object with soil pixels overlain in false color.
}
\description{
Determine the total area (in pixels) of soil present within an image after masking. Masking can be conducted interactively, programmatically with circular or rectangular areas, or automatically using the normalized difference between red and green bands.
}
\details{
\code{area_from_image()} can be called directly to classify soil in a single image, given the path to the image file, or via \code{\link[=process_petri]{process_petri()}} to classify a set of images corresponding to a measurement sequence for a single replicate.

Proper masking of non-soil objects is critical for reliable classification with Otsu's histogram thresholding method (see \code{\link[EBImage:otsu]{EBImage::otsu()}}). Programmatic circular masking (interactive = F, circular = T) is recommended for most use cases. Classification results are provided as a binary image, with each soil object labeled using \code{\link[EBImage:bwlabel]{EBImage::bwlabel()}}, and as a false color overlay to support evaluation of potential sensitive to masking parameters.
}
\examples{
# path to initial (pre-submersion) image
f_init <- system.file("images/exp1_1", "IMG_20220701_120110.jpg", package="SlakeItEasy")

# path to second image (time zero of slaking)
f_t0 <- system.file("images/exp1_1", "IMG_20220701_120230.jpg", package="SlakeItEasy")

# path to final image (ten minutes of slaking)
f_t10 <- system.file("images/exp1_1", "IMG_20220701_121230.jpg", package="SlakeItEasy")

# processing with default circular mask
a_init <- area_from_image(f_init, aggregates_in_initial = 3)
a_t0 <- area_from_image(f_t0)
a_t10 <- area_from_image(f_t10)

cat('initial area: ', a_init$area, '\n',
'area upon submersion: ', a_t0$area, '\n',
'final area: ', a_t10$area, '\n',
'stab10: ', a_init$area/a_t10$area)

par(mfrow = c(1, 3))
plot(a_init$false_color)
plot(a_t0$false_color)
plot(a_t10$false_color)


}
