% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_process.R
\name{batch_process}
\alias{batch_process}
\title{Process a Directory of Replicates}
\usage{
batch_process(
  dir_vec,
  parallel = F,
  ncores = parallel::detectCores() - 1,
  return_output = T,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Process a Directory of Replicates
}
