% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesys_setup.R
\name{check_replicates}
\alias{check_replicates}
\title{Inspect Image Metadata}
\usage{
check_replicates(
  metadata,
  final_img_time_min = 10,
  final_img_tol_sec = 30,
  n_images_min = 3,
  n_images_max = 3
)
}
\arguments{
\item{metadata}{data frame of image metadata as returned by \code{\link[=get_metadata]{get_metadata()}}}

\item{final_img_time_min}{time of final image capture (in minutes) after submersion}

\item{final_img_tol_sec}{tolerance for time of final image capture (in seconds)}

\item{n_images_min}{minimum number of images required per replicate}

\item{n_images_max}{maximum number of images allowed per replicate}
}
\value{
A list of varying length. The first element is a data frame with one row summarizing the evaluation of each replicate. The second element is a character vector of absolute paths to usable replicates. Additional character vectors may be included listing replicates that were flagged for missing images (fewer images than \code{n_images_min}), for having extra images (more images than \code{n_images_max}), for containing images of more than one size, or for lacking a final image at the expected time (\code{final_img_time_min} \eqn{\pm} \code{final_img_tol_sec}).
}
\description{
Evaluate image metadata and identify replicates that do not meet expectations for data analysis.
}
\details{
\code{check_replicates()} evaluates the suitability of replicates (i.e., subdirectories for which metadata were extracted using \code{\link[=get_metadata]{get_metadata()}}) for analysis. Replicates are evaluated based on number of images per subdirectory, image size, and the timing of the final image.
}
\examples{
# Interactively set paths
paths <- set_paths()

# Get image metadata
metadata <- get_metadata(paths$image_dir)

# Inspect image metadata
check <- check_replicates(metadata)


}
