% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_functions.R
\name{crop_image}
\alias{crop_image}
\title{Crop or Mask an Image}
\usage{
crop_image(img, dim_list, mask_only = F)
}
\arguments{
\item{img}{\code{Image} object}

\item{dim_list}{List containing the crop/mask extent as two vectors of length two (minimum and maximum coordinates in the horizontal and in the vertical dimensions)}

\item{mask_only}{Logical. If \code{TRUE}, the output image will retain the dimensions of the input image.}
}
\value{
An object of class \code{Image}.
}
\description{
Produce a rectangular subset of an image, either with the same dimensions as the original image but pixels outside of the subset set to black (mask) or with dimensions aligned to the subset (crop).
}
\details{
\verb{crop image} crops or masks an \code{Image} object based on a rectangular extent provided as a list.

The \code{Image} object returned by \verb{crop image} has attributes \code{crop_dim} and \code{source_dim} with the crop/mask extent and original image dimensions, respectively.
}
\examples{
# path to example image file
f <- system.file("images/exp1_1", "IMG_20220701_120110.jpg", package="SlakeItEasy")

# read in image
img_in <- read_to_portrait(f)

# define extent for crop
e <- list(x = c(115, 540), y = c(220, 640))

# crop image
img_cropped <- crop_image(img_in, e)

EBImage::display(img_cropped)

# mask image
img_masked <- crop_image(img_in, e, mask_only = T)

EBImage::display(img_masked)

}
