% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_funcs.R
\name{cv}
\alias{cv}
\title{Coefficient of Variation}
\usage{
cv(x, na.rm = T)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{logical indicating whether NA values should be excluded from computation}
}
\value{
Coefficient of variation as a percentage of the arithmetic mean of \code{x}
}
\description{
Coefficient of Variation
}
\examples{
# simulate vector of gamma-distributed slaking values
stab10_sim <- 1 / (rgamma(3, 2, 1) + 1)

cv(stab10_sim)

}
