% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesys_setup.R
\name{dir_create_silent}
\alias{dir_create_silent}
\title{Silently Create a Directory}
\usage{
dir_create_silent(direc, ...)
}
\arguments{
\item{direc}{character string containing path name}

\item{...}{additional arguments passed to \code{dir.create}}
}
\description{
Create a directory; suppress warning if directory already exists.
}
