% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesys_setup.R
\name{dir_setup}
\alias{dir_setup}
\title{Create Directory for Output}
\usage{
dir_setup(
  output_dir,
  subdirs = c("to_reprocess", "unusable"),
  return_paths = T
)
}
\arguments{
\item{output_dir}{character string with absolute path to output directory}

\item{subdirs}{character vector with names of subdirectories to create}

\item{return_paths}{logical. If TRUE, a vector containing absolute paths to created subdirectories is returned.}
}
\value{
Character vector with absolute paths to created subdirectories if \code{return_paths} is TRUE.
}
\description{
Create a directory for output of the analysis of a batch of images
}
\details{
\code{dir_setup} creates a batch-level directory for processed images and metadata. Subdirectories not created by \code{\link[=process_petri]{process_petri()}} can be created with \code{dir_setup} by providing a vector of subdirectory names (as relative paths) to the \code{subdirs} argument. By default, subdirectories are created for later manual sorting of images that need reprocessing and images that are unusable for analysis.
}
\examples{
# Interactively set paths
paths <- set_paths()

# Create batch-level directory with default subdirectories to_reprocess and unusable.
# Retain absolute paths to subdirectories as a character vector.
dirs_for_flagged_imgs <- dir_setup(paths$output_dir)
}
