% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesys_setup.R
\name{get_metadata}
\alias{get_metadata}
\title{Extract Image Metadata}
\usage{
get_metadata(
  dir,
  filename_prefix,
  image_extension = "jpg",
  datetime_fmt = "\%Y\%m\%d_\%H\%M\%S"
)
}
\arguments{
\item{dir}{character string with relative or absolute path to directory containing images}

\item{filename_prefix}{character string indicating any text preceding the date and time in image file names}

\item{image_extension}{character string indicating the image file type}

\item{datetime_fmt}{character string indicating the date and time format used in image file names (see \code{\link[base:strptime]{base::strptime()}})}
}
\value{
A data frame (tibble) containing EXIF metadata and temporal information extracted from image file names.
}
\description{
Get EXIF metadata and temporal information for a directory of images.
}
\details{
A simple wrapper around \code{\link[exifr:read_exif]{exifr::read_exif()}}, supplementing EXIF metadata with temporal information extracted from image file names. Within the batch-level directory \code{dir}, it is assumed that images are organized into replicate-level subdirectories. Each subdirectory should contain an initial image of air-dry soil, an image of soil collected upon submersion in water, and a final image collected upon conclusion of the slaking measurement sequence. \code{get_metadata} calculates the elapsed time for each subdirectory as well as the elapsed time of slaking, assuming the second image corresponds to time zero.
}
\examples{
# Interactively set paths
paths <- set_paths(parent_dir_img = 'inst/images')

# Get image metadata
(metadata <- get_metadata(paths$image_dir), filename_prefix = 'IMG_')


}
