% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_functions.R
\name{mask_image_circular}
\alias{mask_image_circular}
\title{Apply a Circular Mask to an Image}
\usage{
mask_image_circular(
  img,
  d = 0.7,
  stepsize_center = 0.01,
  stepsize_diam = 0.01,
  interactive = T,
  h_offset = 0,
  v_offset = 0
)
}
\arguments{
\item{img}{\code{Image} object}

\item{d}{diameter of circular mask as a proportion of image width}

\item{stepsize_center}{step size to reposition center of circular corp}

\item{stepsize_diam}{step size to adjust diameter of circular crop}

\item{interactive}{logical indicating whether the mask should be set interactively}

\item{h_offset}{horizontal offset for center of circular crop}

\item{v_offset}{vertical offset for center of circular crop}
}
\value{
\code{Image} object
}
\description{
Mask pixels outside of a circular area of interest.
}
\details{
For most implementations of the Soil Health Institute's multi-sample aggregate stability standard operating procedure, \code{mask_image_circular} will be the primary image preparation step prior to pixel classification with \code{\link[=rgb_to_binary]{rgb_to_binary()}}.
}
\examples{
# path to example image file
f <- system.file("images/exp1_1", "IMG_20220701_120110.jpg", package="SlakeItEasy")

# read in image
img_in <- read_to_portrait(f)

# mask image

img_masked <- mask_image_circular(img_in, interactive = F)

plot(img_masked)


}
