% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_funcs.R
\name{mean_geom}
\alias{mean_geom}
\title{Geometric Mean}
\usage{
mean_geom(x, na.rm = T)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{logical indicating whether NA values should be excluded from computation}
}
\value{
Geometric mean of \code{x}
}
\description{
Geometric Mean
}
\examples{
# simulate vector of gamma-distributed slaking values
stab10_sim <- 1 / (rgamma(3, 2, 1) + 1)

# geometric mean
mean_geom(stab10_sim)

# arithmetic mean
mean(stab10_sim)

}
