% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_functions.R
\name{norm_diff}
\alias{norm_diff}
\title{Threshold an Image by Normalized Difference between Two Bands}
\usage{
norm_diff(img, b1_idx, b2_idx, mindiff = 0)
}
\arguments{
\item{img}{\code{Image} object}

\item{b1_idx}{index for first band}

\item{b2_idx}{index for second band}

\item{mindiff}{threshold minimum value for normalized difference between bands}
}
\value{
One-band \verb{Image object}, with pixel values of 1 indicating normalized difference values greater than or equal to \code{mindiff} and values of 0 indicating normalize difference values less than \code{mindiff}
}
\description{
Calculate the normalized difference ((x - y) / (x + y)) between two image bands, and segment the image according to a specified value of the normalized difference.
}
\details{
Conservative method of identifying soil within an image. This algorithm is provided to support the masking process when a simple circular mask is insufficient (e.g., non-soil objects are interspersed with soil). \code{norm_diff} will generally underestimate the total area of soil and is intended to be used in conjunction with \code{\link[EBImage:erode_and_dilate]{EBImage::erode_and_dilate()}}.
}
\examples{
# Path to example image
f <- system.file("images/exp1_1", "IMG_20220701_120110.jpg", package="SlakeItEasy")

# Read in image
img <- read_to_portrait(f)

# Apply circular mask to image
masked <- mask_image_circular(img, interactive = F)

# Threshold with normalized difference
nd <- norm_diff(masked, 2, 3)
plot(nd)

}
