% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesys_setup.R
\name{opendirs}
\alias{opendirs}
\title{Open File Explorer}
\usage{
opendirs(dirs)
}
\arguments{
\item{dirs}{character vector with absolute paths to one or more directories to open.}
}
\description{
Open one or more directories in the system file explorer.
}
\examples{
# Interactively set paths
paths <- set_paths()

# Create batch-level directory with default subdirectories to_reprocess and unusable.
# Retain absolute paths to subdirectories as a character vector.
dirs_for_flagged_imgs <- dir_setup(paths$output_dir)

# Open directories
opendirs(dirs_for_flagged_imgs)

}
