% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_petri.R
\name{process_petri}
\alias{process_petri}
\title{Get Soil Area for an Image Sequence}
\usage{
process_petri(
  path_to_image_set,
  outdir,
  filename_prefix,
  image_extension = "jpg",
  datetime_fmt = "\%Y\%m\%d_\%H\%M\%S",
  final_img_time_min = 10,
  final_img_tol_sec = 30,
  match_resolution = T,
  safe_sep = "_x_",
  interactive = F,
  circular_mask = T,
  false_color = "red",
  d = 0.7,
  h_offset = 0,
  v_offset = 0,
  aggregates_in_initial = 3,
  max_rel_dist_from_centroid = 1.3,
  max_rel_dist_from_center = 0.4,
  fixed_crop_fraction_initial = NULL,
  fixed_crop_fraction_run = fixed_crop_fraction_initial,
  max_rel_objlength = 0.4,
  automask_buffer = 0.15,
  erode_kern = 5,
  dilate_kern = 31,
  normdiff_min = 0
)
}
\arguments{
\item{datetime_fmt}{character string indicating the date and time format used in image file names (see \code{\link[base:strptime]{base::strptime()}})}

\item{final_img_time_min}{time of final image capture (in minutes) after submersion}

\item{final_img_tol_sec}{tolerance for time of final image capture (in seconds)}

\item{normdiff_min}{}
}
\value{
Value of the slaking index at \code{final_img_time_min} \eqn{\pm} \code{final_img_tol_sec}
}
\description{
Get Soil Area for an Image Sequence
}
