% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_functions.R
\name{rgb_to_binary}
\alias{rgb_to_binary}
\title{Classify Soil and Background Pixels in a Color Image}
\usage{
rgb_to_binary(img)
}
\arguments{
\item{img}{Three-band \code{Image} object}
}
\value{
One-band \verb{Image object}, with pixel values of 1 indicating soil and 0 indicating background.
}
\description{
Average pixel intensities across red, green, and blue bands, and apply Otsu's binary thresholding algorithm to the pixel intensity histogram.
}
\details{
Pixel intensities are averaged across red, green, and blue bands. The resulting histogram is then partitioned into two classes using Otsu's method for threshold optimization (Otsu 1979). Note that bandwise classification requires 10x computation time and yields results comparable to the average-then-classify approach implemented in {SlakeItEasy}. Because Otsu's method may fail to discern soil from other dark objects within the image, \code{rgb_to_binary} should be applied after cropping/masking the image.
}
\examples{
# path to example image file
f <- system.file("images/exp1_1", "IMG_20220701_120110.jpg", package="SlakeItEasy")

# read in image
img <- read_to_portrait(f)

# apply circular mask to image
img_masked <- mask_image_circular(img, interactive = F)

classified <- rgb_to_binary(img_masked)

par(mfrow = c(1, 2))
plot(img); plot(classified)

}
\references{
N. Otsu. A threshold selection method from gray-level histograms. IEEE Trans. Sys., Man., Cyber. 9 (1): 62-66, 1979. doi:10.1109/TSMC.1979.4310076.
}
