% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_functions.R
\name{set_crop}
\alias{set_crop}
\title{Interactively Determine an Extent for Rectangular Cropping}
\usage{
set_crop(npts = 2)
}
\arguments{
\item{npts}{Number of points for extent}
}
\value{
Two-element list with first element a vector of minimum and maximum horizontal coordinates and second element a vector of minimum and maximum vertical coordinates.
}
\description{
Interactively Determine an Extent for Rectangular Cropping
}
\examples{
# path to example image file
f <- system.file("images/exp1_1", "IMG_20220701_120110.jpg", package="SlakeItEasy")

# read in image
img_portrait <- read_to_portrait(f)

# plot image
plot(img_portrait)

# get extent for crop
(ext <- set_crop())

}
