% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlakeItEasy-package.R
\docType{package}
\name{SlakeItEasy-package}
\alias{SlakeItEasy}
\alias{SlakeItEasy-package}
\title{SlakeItEasy: Image-Based Estimation of Soil Aggregate Stability}
\description{
Utility functions to determine the projected area of soil during slaking tests, following Fajardo et al. (2016) \doi{10.1016/j.still.2016.05.018}. Functions are designed for output of high-throughput image aquisition schemes (e.g., multiple replicates, multiple samples).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://soilhealthinstitute.org/}
  \item \url{https://github.com/Soil-Health-Institute}
}

}
\keyword{internal}
